remove client a2a system message which is not needed & fix body maintenance flow

+++ section-2/step-05/remote-a2a-agent/src/main/java/com/demo/DispositionAgent.java
@@ -8,6 +8,10 @@ import jakarta.enterprise.context.ApplicationScoped;
 
 /**
  * Agent that determines how to dispose of a car based on value, condition, and damage.
+ *
+ * Note: Uses Integer for carNumber (instead of Long used in step-04) because the A2A protocol
+ * requires explicit string-to-type parsing in DispositionAgentExecutor. Local agents in step-04
+ * can use Long as LangChain4j handles type conversion automatically.
  */
 @RegisterAiService
 @ApplicationScoped
@@ -49,7 +53,7 @@ public interface DispositionAgent {
             String carMake,
             String carModel,
             Integer carYear,
-            Long carNumber,
+            Integer carNumber,
             String carCondition,
             String carValue,
             String rentalFeedback);
